+++ R/r_squared.R
@@ -1,13 +1,14 @@
-#' Pseudo R-Squared
+#' Generalized R-Squared
 #'
 #' Returns (weighted) proportion of deviance explained, see reference below. For the mean-squared error as deviance, this equals the usual (weighted) R-squared. The higher, the better.
 #'
-#' For simplicity, the deviance gain is calculated regarding the null model derived from the actual values.
+#' The deviance gain is calculated regarding the null model derived from the actual values. While fine for in-sample considerations, this is only an approximation for out-of-sample considerations. There, it is recommended to calculate null deviance regarding the in-sample (weighted) mean. This value can be passed by the argument \code{reference_mean}.
 #'
 #' @param actual Observed values.
 #' @param predicted Predicted values.
 #' @param w Optional case weights.
 #' @param deviance_function A positive (deviance) function taking four arguments: "actual", "predicted", "w" and "...".
+#' @param reference_mean An optional reference mean used to derive the null deviance. Recommended in out-of-sample applications.
 #' @param ... Further arguments passed to \code{weighted_mean} and \code{deviance_function}.
 #' @return A numeric vector of length one.
 #' @export
@@ -29,9 +30,15 @@
 #' r_squared(1:10, c(1, 1:9), deviance_function = myTweedie)
 #' @seealso \code{\link{deviance_normal}, \link{mse}}.
 r_squared <- function(actual, predicted, w = NULL,
-                      deviance_function = mse, ...) {
+                      deviance_function = mse,
+                      reference_mean = NULL, ...) {
   stopifnot(is.function(deviance_function))
-  actual_mean <- rep(weighted_mean(actual, w = w, ...), length(actual))
+  if (is.null(reference_mean)) {
+    actual_mean <- rep(weighted_mean(actual, w = w, ...), length(actual))
+  } else {
+    stopifnot(length(reference_mean) == 1L)
+    actual_mean <- rep(reference_mean, length(actual))
+  }
   d0 <- deviance_function(actual = actual, predicted = actual_mean, w = w, ...)
   dres <- deviance_function(actual = actual, predicted = predicted, w = w, ...)
   1 - dres / d0